Add animated video
