+++ src/components/NavItems/tools/GIF/Hooks/useGetGif.js
@@ -7,7 +7,7 @@ import { setStateDownloading,setStateBackResults } from "../../../../../redux/ac
 import { saveAs } from 'file-saver';
 import useAuthenticatedRequest from "../../../../Shared/Authentication/useAuthenticatedRequest";
 
-const useGetGif = (images, delayInput, enableDownload) => {
+const useGetGif = (images, delayInput, enableDownload, downloadType) => {
     const keyword = useLoadLanguage("components/NavItems/tools/Forensic.tsv", tsv);
     const dispatch = useDispatch();
     const authenticatedRequest = useAuthenticatedRequest();
@@ -24,10 +24,25 @@ const useGetGif = (images, delayInput, enableDownload) => {
         };
 
 
-        const downloadGif = (response) => {
-            //console.log(response);
-            const file = new Blob([response.data], { type: 'image/gif' });
-            saveAs(file, "image.gif");
+        const download = (response) => {
+            let contentType = response.headers["content-type"]
+            //console.log("header content-type ", contentType)
+            const content = new Blob([response.data], { type: contentType });
+            switch(contentType)
+            {
+                case 'video/mp4':
+                    var id = response.headers.videoid
+                    saveAs(content, "vera-animated-video-"+id+".mp4");
+                break;
+                case 'image/gif':
+                    var id_gif = response.headers.imageid
+                    saveAs(content, "vera-animated-image-"+id_gif+".gif");
+                break;
+                default:
+                    var id_gif_def = response.headers.imageid
+                    saveAs(content, "vera-animated-image-"+id_gif_def+".gif");
+                    break;
+            }
             dispatch(setStateBackResults());
         }
 
@@ -44,25 +59,24 @@ const useGetGif = (images, delayInput, enableDownload) => {
                 ],
                 delay: delayInput
             }
-
+            let endpoint = downloadType === "mp4" ? "/video" :"/animated"
             const axiosConfig = {
                 method: "post",
-                url: baseURL + "/animated",
+                url: baseURL + endpoint,
                 data: body,
                 responseType: 'blob',
             }
 
             authenticatedRequest(axiosConfig)
-                .then(response => downloadGif(response))
+                .then(response => {
+                    download(response)
+                })
                 .catch(error => {
                     handleError("gif_error_" + error.status);
                 });
-            
-
         };
 
-
-
-    }, [images, delayInput, enableDownload, keyword, dispatch, baseURL, authenticatedRequest]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [images, delayInput, enableDownload, downloadType, authenticatedRequest]);
 };
 export default useGetGif;
